upload code of price issue fixed

+++ app/Models/Product.php
@@ -61,7 +61,8 @@ class Product extends Model
 
     public function getPriceAttribute($value)
     {
-        return (int) $value;
+        // return (int) $value;
+         return (double) $value;
     }
 
     public function getIsFavouriteAttribute($value)
